+++ arpg-timeline/src/components/Dashboard/CantFindGame.tsx
@@ -1,0 +1,10 @@
+export const CantFindGame = () => (
+    <div className="bg-card 3xl:nth-[2]:order-first 3xl:nth-[3]:order-first 3xl:nth-[4]:order-first 4xl:nth-[2]:order-first 4xl:nth-[3]:order-first 4xl:nth-[4]:order-first 4xl:nth-[5]:order-first order-last flex flex-col items-center justify-center rounded-lg p-6 text-center md:nth-[2]:order-first lg:nth-[2]:order-first xl:nth-[2]:order-first">
+        <h3 className="mb-2 text-lg font-semibold text-gray-900 dark:text-white">
+            Can&apos;t find a game?
+        </h3>
+        <p className="text-sm text-gray-600 dark:text-gray-400">
+            See other tabs at the top or request on Discord
+        </p>
+    </div>
+);
