+++ arpg-timeline/src/components/SimpleAnalytics.tsx
@@ -6,6 +6,10 @@ import { sa_event } from "@/lib/sa_event";
 
 export const SimpleAnalytics = () => {
     useEffect(() => {
+        if (typeof document === "undefined") {
+            return;
+        }
+
         const listener = (e: MouseEvent) => {
             const target = e.target as HTMLElement;
             const saClickSource = target.dataset.saClick;
@@ -18,7 +22,11 @@ export const SimpleAnalytics = () => {
 
         document.addEventListener("click", listener);
 
-        return () => document.removeEventListener("click", listener);
+        return () => {
+            if (typeof document !== "undefined") {
+                document.removeEventListener("click", listener);
+            }
+        };
     }, []);
 
     return null;
